test: rename uai tests to avoid file name duplication
