+++ tests/test_raillabel/_understand_ai_t4_format/test_uai_bounding_box_3d.py
