+++ tests/test_raillabel/_understand_ai_t4_format/test_uai_coordinate_system.py
