+++ tests/test_raillabel/_understand_ai_t4_format/test_uai_scene.py
@@ -9,7 +9,7 @@ import pytest
 
 sys.path.insert(1, str(Path(__file__).parent.parent.parent.parent))
 
-from test_frame import _prepare_frame_data
+from test_uai_frame import _prepare_frame_data
 
 import raillabel._understand_ai_t4_format as uai_format
